requirements.txt
allure-pytest==2.13.4
allure-python-commons==2.13.4
async-generator==1.10
attrs==22.2.0
blinker==1.5
Brlapi==0.8.4
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
coverage==7.4.4
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.3.2
distlib==0.3.8
distro==1.8.0
exceptiongroup==1.1.0
Faker==24.11.0
filelock==3.14.0
fuse-python==1.0.5
gpg==1.18.0
gyp==0.1
h11==0.14.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
louis==3.24.0
more-itertools==8.10.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
platformdirs==4.2.2
pluggy==1.0.0+repack
py==1.11.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.45.2
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
pylibacl==0.7.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pytest==7.2.1
pytest-cov==5.0.0
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-debianbts==4.0.1
pyxattr==0.8.1
pyxdg==0.28
reportbug==12.0.0
requests==2.28.1
selenium==4.8.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
virtualenv==20.26.2
wadllib==1.3.6
wsproto==1.2.0
xdg==5
